Refactor ProductStackNavigator
